% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/encoding_check.R
\name{isBIG5}
\alias{isBIG5}
\title{Indicate whether the encoding of input string is BIG5.}
\usage{
isBIG5(string, combine = FALSE)
}
\arguments{
\item{string}{A character vector.}

\item{combine}{Whether to combine all the strings.}
}
\value{
A vector of logical values, the length of \code{string} - or a single logical value if \code{combine}
is TRUE.
}
\description{
Indicate whether the encoding of input string is BIG5.
}
\author{
Jian Li <\email{rweibo@sina.com}>
}

