% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding_check.R
\name{isGB18030}
\alias{isGB18030}
\title{Indicate whether the encoding of input string is GBK.}
\usage{
isGB18030(string, combine = FALSE)
}
\arguments{
\item{string}{A character vector.}

\item{combine}{Whether to combine all the strings.}
}
\value{
A vector of logical values, the length of \code{string} - or a single logical value if \code{combine}
is TRUE.
}
\description{
Indicate whether the encoding of input string is GBK.
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
